+++ features/dd-sdk-android-ndk/src/main/cpp/datadog-ndk.h
