+++ sample/kotlin/src/main/cpp/datadog-ndk-sample.cpp
